+++ include/game-models/AbilityEndowedPlayer/ability-endowed-player.h
@@ -24,7 +24,7 @@ public:
 		PlayerTeamId teamId,
 		int initialHitPoints,
 		int initialAmmo,
-		double damage,
+		int damage,
 		long long abilityCoolDown_ms
 	);
